created a common response model with success, data and error.
Also, the routes_test is wrong. Realised this after seeing that because of common model, have to use interface{} for data, that interface in test's when converting to json comes as a map, so have to convert from map to the struct again for comparison. which kinda felt too much and plainly wrong.
And also, these tests are from the client side. So, these tests should no nothing about server models, they should just test by json